// github/types/paths/gists/{gist_id}.types.ts
export type gistsGet = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsGet_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: gist_simple;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      403: forbidden_gist;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type gistsUpdate = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsUpdate_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The description of the gist.
       * @example "Example Ruby script"
       */
      description?: string;
      /**
       * The gist files to be updated, renamed, or deleted. Each `key` must match the current filename
       * (including extension) of the targeted gist file. For example: `hello.py`.
       *
       * To delete a file, set the whole file to null. For example: `hello.py : null`. The file will also be
       * deleted if the specified object does not contain at least one of `content` or `filename`.
       * @example {"hello.rb":{"content":"blah","filename":"goodbye.rb"}}
       */
      files?: {
        [key: string]: {
          /**
           * The new content of the file.
           */
          content?: string;
          /**
           * The new filename for the file.
           */
          filename?: string;
        };
      };
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: gist_simple;
          };
        };
        examples: {
          updateGist: unknown;
          deleteFile: unknown;
          renameFile: unknown;
        };
      };
      404: not_found;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type gistsDelete = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsDelete_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type gistsGet_Path = { gist_id: string };

export type gistsUpdate_Path = { gist_id: string };

export type gistsDelete_Path = { gist_id: string };